Added: From given 3 float values finding the maximum

+++ TestMaximum.java
@@ -2,28 +2,28 @@
 /*
  * The Generics program implements Test cases that simply
  * finds Maximum of 3 variables using Generics.
- * Master - Given 3 Integers find the maximum
+ * Master - Given 3 Float. find the maximum
  * @author: Lokesh
  * @date: 01-04-2022
  */
 
 package com.BridgeLab.Genrics;
-public  class TestMaximum {
-    int x, y, z;
+public class TestMaximum {
+    float x, y, z;
 
     // constructor
-    public TestMaximum(int x, int y, int z) {
+    public TestMaximum(Float x, Float y, Float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
     /*
-     * testMaximum method created for checking max of three integers
+     * testMaximum method created for checking max of three float values
      */
-    public static Integer testMaximum(Integer x, Integer y, Integer z) {
+    public static Float testMaximum(Float x, Float y, Float z) {
         // Initializing x as greater
-        Integer max = x;
+        Float max = x;
         if (y.compareTo(max) > 0)
             // as of now y is greater
             max = y;
@@ -31,7 +31,7 @@ public  class TestMaximum {
         if (z.compareTo(max) > 0)
             // now z is greater
             max = z;
-        printMax(x, y, z, max);
+        printMax(x,y,z,max);
         return max;
 
     }
@@ -39,7 +39,7 @@ public  class TestMaximum {
     /*
      *  printMax method created for displaying maximum values from the given 3 variables.
      */
-    public static void printMax(Integer x, Integer y, Integer z, Integer max) {
+    public static void printMax(Float x, Float y, Float z, Float max) {
         System.out.printf("The max of %s, %s and %s is %s\n", x, y, z, max);
 
     }
